refactor: move `getImageUrlFromEntityProperties` into shared folder
