+++ apps/hash-frontend/src/pages/[shortname].page/edit-user-profile-info-modal/user-profile-info-modal-header.tsx
@@ -24,7 +24,7 @@ import { User } from "../../../lib/user-and-org";
 import { TrashRegularIcon } from "../../../shared/icons/trash-regular-icon";
 import { XMarkRegularIcon } from "../../../shared/icons/x-mark-regular-icon";
 import { Button, ButtonProps } from "../../../shared/ui";
-import { getImageUrlFromEntityProperties } from "../../[shortname]/entities/[entity-uuid].page/entity-editor/shared/get-image-url-from-properties";
+import { getImageUrlFromEntityProperties } from "../../shared/get-image-url-from-properties";
 import { useAuthInfo } from "../../shared/auth-info-context";
 import { leftColumnWidth } from "../util";
