+++ apps/hash-frontend/src/pages/shared/get-image-url-from-properties.ts
